refactor(automation): Tempo 统一为 Automation 一员
- value 类型 u16→f32 全局统一，序列化迁移
- conductor.tempo 改为 AutomationLane，存于 Arc<ConductorData>
- 删除 decor tempo 渲染路径（tempo.rs），Tempo 合并到 curve pipeline
- AutomationEdit/Delta 添加 target 字段，编辑方法按 target 分派到 conductor.tempo 或 track.automation_lanes
- interaction.rs 支持 Tempo 模式编辑（Add/Move/Delete/CycleShape）
- info_panel.rs 的 snapshot/undo 支持 Tempo
- 删除 AutomationPanelView.show_tempo，Tempo 通过 selected_target == AutomationTarget::Tempo 选择
- AUTOMATION_TARGETS 列表加入 Tempo，popup 菜单简化
- prepare.rs 删除 tempo_lane_hash 和 tempo_lane 参数，cache key 由 target_hash + fixed_lanes_hash 覆盖

+++ crates/yinhe-egui/src/app.rs
@@ -71,7 +71,7 @@ pub struct App {
     pub(crate) right_tab: Option<crate::right_panel::RightTab>,
     pub(crate) info_content: Option<crate::right_panel::InfoContent>,
     /// 拖拽锚点时的 ghost 值（tick, value），供信息面板实时显示
-    pub(crate) automation_drag_ghost: Option<(u32, u16)>,
+    pub(crate) automation_drag_ghost: Option<(u32, f32)>,
 
     // ── Tool palette ──
     pub(crate) active_tool: crate::widgets::tools_panel::Tool,
